Split SSHConfig to own file.
